requirements.txt
smolagents==1.4.1
google-search-results==2.4.2
python-dotenv==1.0.1
firecrawl-py==1.10.0
litellm==1.59.0
